Change name of folder and components
