--- client/src/features/calendar/components/Info/PlantsInfo.tsx
@@ -1,30 +1,0 @@
-import { Plant } from "../../../../types";
-
-interface Props {
-  plants: Plant[];
-  watered?: boolean;
-  handleDate: (plants: Plant[]) => void;
-}
-
-export const PlantsInfo = ({ plants, watered, handleDate }: Props) => {
-  if (plants.length < 1) return null;
-
-  return (
-    <div>
-      <ul className="flex flex-col gap-1 pl-2">
-        {plants.map((plant) => {
-          return (
-            <li key={plant.id} className="">
-              <button
-                className={`text-xl font-medium ${watered && "line-through"}`}
-                onClick={() => handleDate([plant])}
-              >
-                {plant.name}
-              </button>
-            </li>
-          );
-        })}
-      </ul>
-    </div>
-  );
-};
